跟据页、id、分类查询菜品（包括分类名称、菜品风味）

+++ sky-server/src/main/java/com/sky/mapper/DishMapper.java
@@ -17,8 +17,8 @@ public interface DishMapper {
     List<Dish> list(DishPageQueryDTO dishPageQueryDTO);
 
     @Select("select * from dish where category_id=#{categoryId}")
-    List<Dish> getByCategoryId(Integer categoryId);
+    List<Dish> getByCategoryId(Long categoryId);
 
     @Select("select *from dish where id=#{id}")
-    Dish getById(int id);
+    Dish getById(Long id);
 }
